Added due date to to do app
Each task can have a due date. Might change up the styling
based on when the due date is.

Will add a footer for extra brownie points.

In future iterations, will be able to move tasks to
different projects. As well as display what are TODAYs tasks.

Last step will be to allow for local storage.