Completed basic functionality of our to do list.
This was a big commit and we learned a lot.
0. Projects are now tied with tasks. You can now keep context of
different states of the task object as we traverse through different
projects. We added due dates and priority features. The most challenging
thing was dynamically changing a tasks projects while updating the interface.
I was not able to include everything, but hopefully in the future!

1. Building projects is not easy. We need to work on building
components better from the start. Building a small working app is
okay for learning, but not having enough modular pieces is VERY
difficult to deal with. In the future, I will be MUCH more cognizant
about laying out the roadmap of how we want to structure the program.
This was my first BIG project and the lessons were invaluable.

2. We used objects and single purpose objects, but we could have been
MORE strict with what is allowed in a function. I need to calm down
and stop rushing. This isn't a race. Building GOOD sections takes time.

3. I need to commit WAY more frequently. I keep making small changes over
and over but forgetting (or avoiding) to commit.

4. Having said all that I need to move on to the next project. We will
spend one more day learning localStorage, plus some MINOR styling fixes.
We will no longer work on improving the functionality of this project until
the far future where we have better tools to build this type of app.

5. Overall, despite spending nearly 30 hours on this project and it still
lacking functionality (A LOT is there, but some stuff I've disabled), I am
very proud of what I was able to accomplish through self-learning. I can
confidently say that I've expanded my knowledge SIGNIFICANTLY since when
I first started 4 days ago.